requirements.txt
opencv-python
cvzone
face_recognition
numpy
psycopg2-binary
ultralytics
mediapipe